WIP: T05 -- Extract instance-lifecycle test classes from tests/unit/core/test_app_lifecycle_service.py
